Move openfeign out of spring-mvc
