+++ README.md
@@ -176,13 +176,13 @@ If the target application runs under JDK 11 and above, the following arguments s
 | Jersey                      | 1.19.4      |             |                                                | &check; |
 | MongoDB                     | 3.4.2       |             | &check;                                        |         |
 | MySQL                       | 5.x         | 8.x         | &check;                                        |         |
+| Open Feign                  | 10.8        |             |                                                | &check; |
 | Quartz                      | 2.x         |             | &check;                                        | &check; |
 | Redis - Jedis               | 2.9         | 5.x         | &check;                                        | &check; |
 | Redis - Lettuce(2)          | 5.1.2       | 6.x         | &check;                                        | &check; |
 | Redis - Redisson            | 3.19.0      |             | &check;                                        | &check; |
 | Spring Boot                 | 1.5         | 3.0+        |                                                | &check; |
 | Spring Bean                 | 4.3.12      |             |                                                | &check; |
-| Spring Open Feign           | 10.8        |             |                                                | &check; |
 | Spring Rest Template        | 4.3.12      |             |                                                | &check; |
 | Spring Scheduling           | 4.3.12      |             |                                                | &check; |
 | Spring Gateway              | 3.0.0       |             | [&check;](doc/metrics/http-outgoing/README.md) | &check; |
