+++ agent/agent-plugins/openfeign/src/main/java/org/bithon/agent/plugin/openfeign/OpenFeignPlugin.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.agent.plugin.spring.mvc;
+package org.bithon.agent.plugin.openfeign;
 
 import org.bithon.agent.instrumentation.aop.interceptor.descriptor.InterceptorDescriptor;
 import org.bithon.agent.instrumentation.aop.interceptor.plugin.IPlugin;
@@ -27,23 +27,21 @@ import static org.bithon.agent.instrumentation.aop.interceptor.descriptor.Interc
 /**
  * @author frankchen
  */
-public class SpringMvcPlugin implements IPlugin {
+public class OpenFeignPlugin implements IPlugin {
 
     @Override
     public List<InterceptorDescriptor> getInterceptors() {
 
         return Arrays.asList(
-
-            forClass("org.springframework.web.client.RestTemplate")
-                .onMethod("doExecute")
-                .interceptedBy("org.bithon.agent.plugin.spring.mvc.RestTemplate$Execute")
-                .onMethod("handleResponse")
-                .interceptedBy("org.bithon.agent.plugin.spring.mvc.RestTemplate$HandleResponse")
+            forClass("feign.SynchronousMethodHandler$Factory")
+                .onMethod("create")
+                .andArgs(1, "feign.MethodMetadata")
+                .interceptedBy("org.bithon.agent.plugin.openfeign.interceptor.SynchronousMethodHandlerFactory$Create")
                 .build(),
 
-            forClass("org.springframework.web.method.support.InvocableHandlerMethod")
-                .onMethod("doInvoke")
-                .interceptedBy("org.bithon.agent.plugin.spring.mvc.InvocableHandlerMethod$DoInvoke")
+            forClass("feign.SynchronousMethodHandler")
+                .onMethod("invoke")
+                .interceptedBy("org.bithon.agent.plugin.openfeign.interceptor.SynchronousMethodHandler$Invoke")
                 .build()
         );
     }
