+++ agent/agent-plugins/pom.xml
@@ -37,6 +37,7 @@
     <module>httpclient-jetty</module>
     <module>httpclient-netty3</module>
     <module>httpclient-reactor</module>
+    <module>openfeign</module>
     <module>redis-jedis-2.x</module>
     <module>redis-jedis-3.x</module>
     <module>redis-jedis-4.x</module>
@@ -248,6 +249,13 @@
     <dependency>
       <groupId>org.bithon.agent</groupId>
       <artifactId>agent-plugin-logging-log4j2</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+
+    <!-- Open Feign -->
+    <dependency>
+      <groupId>org.bithon.agent</groupId>
+      <artifactId>agent-plugin-openfeign</artifactId>
       <version>${project.version}</version>
     </dependency>
